+++ utilities/projen/index.ts
@@ -1,6 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+export * from './lib/build';
 export * from './lib/deps';
 export * from './lib/generate';
 export * from './lib/ignore';
