+++ utilities/projen/lib/ignore.ts
@@ -1,12 +1,12 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { project } from '../../../.projenrc';
+import { CdklabsTypeScriptProject } from 'cdklabs-projen-project-types';
 
 /**
  * Sets up the default ignores for the project.
  */
-export const configureIgnores = () => {
+export const configureIgnores = (project: CdklabsTypeScriptProject) => {
     const ignores = [
         '.DS_Store',
         '.python-version',
